% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_Gibbs_samplers.R
\name{add_storage}
\alias{add_storage}
\title{Add storage space to an array}
\usage{
add_storage(x, n)
}
\arguments{
\item{x}{An array.}

\item{n}{An integer.}
}
\value{
\code{x} with \code{n} added 'rows' in the first dimension.
}
\description{
Add storage space to an array
}
\examples{
a <- array(1:24, dim = c(2, 3, 4))
BDSE:::add_storage(a, 2)
}
