% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{assign_first}
\alias{assign_first}
\title{Assign to first element in first dimension of an array}
\usage{
assign_first(arr, arr1)
}
\arguments{
\item{arr}{Array to assign values to.}

\item{arr1}{A vector, matrix or array containing values to assign to the
first element/dimension of \code{arr}.}
}
\value{
Invisibly returns \code{arr1}.
}
\description{
Intended behaviour: \code{arr[1, ...] <- arr1}.
}
\details{
\code{arr} and \code{arr1} must exist in the parent environment.
}
\examples{
# Example 1:
a <- array(1, dim = 5)
a_1 <- 0
BDSE:::assign_first(a, a_1)
a

# Example 2:
a <- array(1:12, dim = c(3, 2, 2))
a_1 <- matrix(rep(100, 4), nrow = 2)
BDSE:::assign_first(a, a_1)
a
}
