% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dataset}
\alias{base_dataset}
\title{Synthetic base dataset}
\format{
A data.frame with 37108 rows and 20 variables:
\describe{
  \item{ethnicity}{...}
  \item{region}{...}
  \item{sex}{...}
  \item{age_grp}{...}
  \item{qualification}{...}
  \item{occupation}{...}
  \item{hours}{...}
  \item{income}{...}
  \item{hours_cen_mean}{...}
  \item{hours_cen_40hrs}{...}
  \item{hours_grp}{...}
  \item{age}{...}
  \item{region_fact}{...}
  \item{ethnicity1}{...}
  \item{ethnicity2}{...}
  \item{ethnicity3}{...}
  \item{ethnicity4}{...}
  \item{ethnicity5}{...}
  \item{ethnicity6}{...}
  \item{ethnicity9}{...}
}
}
\usage{
base_dataset
}
\description{
This is an extension of \code{base_dataset_raw}. It includes individuals with
ages below 15, single-year ages for al records and ethnicity indicators.
}
\keyword{datasets}
