% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dataset_raw}
\alias{base_dataset_raw}
\title{Base dataset (From StatsNZ Website)}
\format{
A data.frame with 29471 rows and 8 variables:
\describe{
  \item{ethnicity}{...}
  \item{region}{...}
  \item{sex}{...} 
  \item{age_grp}{...}
  \item{qualification}{...}
  \item{occupation}{...}
  \item{hours}{...}
  \item{income}{...}
}
}
\usage{
base_dataset_raw
}
\description{
A dataset containing covariate values, which can be used to generate a 
realistic synthetic target population dataset.
}
\keyword{datasets}
