% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{bind_chains}
\alias{bind_chains}
\title{Binds elements of lists}
\usage{
bind_chains(results_list, pars)
}
\arguments{
\item{results_list}{A list of lists (i.e. the list of chains from the Gibbs
sampler).}

\item{pars}{A character vector specifying which groups of parameters must be
retained (e.g. "n_target", "coefficients1", "cluster_sds1").}
}
\value{
A list of matrices, where each element of the list coresponds to a 
single chain.
}
\description{
Given a list of sampler chain outputs, this function binds the elements of	
each list (i.e. chain) into a single matrix, and returns a list of matrices,	
where each matrix corresponds to a chain output.
}
