% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate credible intervals}
\usage{
calc_ci(
  posterior_sample,
  true_values = NULL,
  mle_values = NULL,
  ci_coverage = 0.9,
  ci_method = c("ETI", "HDI"),
  ci_probs = NULL
)
}
\arguments{
\item{posterior_sample}{A data.frame of posterior samples.}

\item{true_values}{An optional numeric vector specifying true parameter
values.}

\item{ci_coverage}{A scalar in the interval \eqn{[0, 1]} specifying the
coverage of the CI.}

\item{ci_method}{A string (either "ETI" or "HDI") specifying the CI method to
be used.}

\item{ci_probs}{A two-element numeric vector specifying the lower and upper
probabilities for the credible intervals. (Defaults to \code{NULL}).}

\item{mle_value}{An optional numeric vector specifying the maximum likelihood
estimates.}
}
\value{
A data.frame, where rows correspond to parameters, and the columns
  are the lower and higher bound of the CI, the posterior median and the true
  values and ML estimates (if provided as input).
}
\description{
Calculates either the equal-tail (ETI) or highest-density (HDI) posterior
intervals using the credible interval (CI) function
\code{\link[bayestestR]{ci}}. An option is provided here to manually set
probabilities and get the desired quantiles for the CI. (This option is
equivalent to choosing the CI method "ETI", but median values might vary
slightly due to implementation here.)
}
\details{
If \code{ci_probs} is specified, it is used to get the quantiles for the CI;
otherwise, \code{ci_coverage} and \code{ci_method} will be used.
}
