% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_likelihoods.R
\name{calc_loglike_binom_n_trial}
\alias{calc_loglike_binom_n_trial}
\title{Calculate log-likelihood of binomial number of trials}
\usage{
calc_loglike_binom_n_trial(N, k, p)
}
\arguments{
\item{N}{An integer specifying the number of trials.}

\item{k}{An integer specifying the number of successes.}

\item{p}{A numeric value in \eqn{(0, 1)} specifying the probability of
success in each trial.}
}
\value{
A numeric value (the logarithm of the likelihood).
}
\description{
This helper function is used in \code{\link[BDSE]{draw_n_target_uniform}}, to
calculate the log-likelihood of the unknown number of trials \code{N} of a
binomial model, for known number of successes and probability of success.
}
\examples{
calc_loglike_binom_n_trial(N = 250,
                           k = 200,
                           p = 0.6)

}
