% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_likelihoods.R
\name{calc_loglike_logistic}
\alias{calc_loglike_logistic}
\title{Calculate log-likelihood of a logistic regression model}
\usage{
calc_loglike_logistic(
  X,
  y,
  coefficients,
  clusters = NULL,
  n_clusters = NULL,
  offsets = 0,
  weights = 1,
  counts = 1
)
}
\arguments{
\item{X}{A numeric (design) matrix with dimension \eqn{n} x \eqn{p}, where
\code{n} is the number of individuals, and \eqn{p} is the number of
covariates (including the intercept).}

\item{y}{A binary (0, 1) vector of length \eqn{n}, representing observations
of the response variable.}

\item{coefficients}{A numeric vector of length \eqn{p + 1} of
individual-level effects in the logistic regression model, the first
element of which is the intercept.}

\item{clusters}{A vector of length \eqn{n}, with values in
\code{1:n_clusters}, representing the cluster membership of the individual
observations.}

\item{n_clusters}{An integer specifying the number of unique clusters in the
population, some of which may not be represented in the observed data.}

\item{offsets}{An optional numeric vector of length \eqn{n} specifying the
values added to the linear component to account for effects not dependent
on \code{X}.}

\item{weights}{An optional numeric vector of length \eqn{n} specifying the
sample weights.}

\item{counts}{A numeric vector of counts of the number of observations in
each corresponding row of the design matrix \code{X}.}
}
\value{
If \code{is.null(clusters)}, a scalar log-likelihood. If clusters
are provided, a named vector of length \code{n_clusters} of log-likelihoods
by cluster.
}
\description{
This function calculates the log-likelihood for a vector of observed
Bernoulli indicators, for a logistic regression model:
\deqn{
   [Y | \beta_0, \beta_1, X = x] ~ Bernoulli(p(x; \beta_0, \beta_1)) ,
}
where \eqn{\beta_0} is a group-level effect and \eqn{\beta_1} is a
vector of individual-level effects. The logistic model is
\deqn{
  logit Pr(Y = 1 | X = x, \beta_0, \beta_1) = \beta_0 + x^T \beta_1 .
}
}
\details{
In \code{calc_loglike_logistic}, \code{coefficients} is the vector of
individual-level effects \eqn{\beta_1}, and
\code{offsets} is the vector of group-level (or cluster-level) effects
\eqn{\beta_0}.

If \code{clusters} is given, then the log-likelihood values are calculated
for each cluster.

If the data does not include observations from all clusters, i.e. the
number of unique values in the vector of observed clusters \code{clusters}
is less than \code{n_clusters}, then the log-likelihood for the unobserved
clusters is set to zero.
}
\examples{
n            <- 100   # number of observations
n_clusters   <- 5    # number of clusters
clusters     <- sample(n_clusters - 1, n, replace = TRUE)

X            <- cbind('(Intercept)' = rep(1, n),
                      age  = runif(n, 1, 85),
                      sexM = rbinom(n, 1, 0.8))
coefficients <- c('(Intercept)' = 0.2,
                  age           = 0.01,
                  sexM          = 0.75)
y            <- rbinom(n, 1, 0.03)
offsets      <- rnorm(n_clusters, 0, 1)[clusters]
weights      <- sample(3, n, replace = TRUE)

ll1 <- BDSE:::calc_loglike_logistic(X            = X,
                                    y            = y,
                                    coefficients = coefficients, 
                                    clusters     = clusters,
                                    n_clusters   = n_clusters,
                                    offsets      = offsets,
                                    weights      = weights)
ll1

ll2 <- BDSE:::calc_loglike_logistic(X            = X,
                                    y            = y,
                                    coefficients = coefficients, 
                                    offsets      = offsets,
                                    weights      = weights)

ll2

sum(ll1) == ll2

}
