% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_missing_probabilities}
\alias{calc_missing_probabilities}
\title{Calculate list exclusion probabilities}
\usage{
calc_missing_probabilities(...)
}
\arguments{
\item{...}{Named lists with elements: \code{X} (a design matrix of dimension
\eqn{n x p}); \code{coefficients} (a \eqn{p}-dimensional vector of
individual-level effects); and \code{offsets} (a vector of offsets, e.g.
group-level effects for individuals).}
}
\value{
A vector of probabilities.
}
\description{
Calculates the probability of being missed by all lists, given the observed
covariates data.
}
\details{
The probability of not being captured by either of the two lists is \deqn{
P{L_1 = 0, L_2 = 0 | X = x, \phi_1, \phi_2} = P{L_1 = 0| X = x, \phi_1} *
P{L_2 = 0 | X = x, \phi_2} } assuming that, conditional on the covariate
values, inclusion on List 1 and List 2 are independent. For list \eqn{j},
\eqn{\phi_j} is the vector of parameters, and includes individual-level
effects (\code{coefficients}) and group-level effects (which make up the
\code{offsets}).
}
\examples{
n          <- 1000  # number of observations
n_clusters <- 10    # number of clusters
clusters   <- sample(n_clusters, n, replace = TRUE)

X          <- cbind('(Intercept)' = rep(1, n),
                    sexM          = rbinom(n, 1, 0.5),
                    age           = sample(85, n, replace = TRUE))
coeffs1    <- c(0.5, -0.75, 0.01)
coeffs2    <- c(0.5, 0.9, 0.05)
effects1   <- rnorm(n_clusters, 0, 1)
effects2   <- rnorm(n_clusters, 0, 0.5) 
offsets1   <- effects1[clusters]
offsets2   <- effects2[clusters]

## Example 1:
p_mis <- BDSE:::calc_missing_probabilities(
           list(X            = X,
                coefficients = coeffs1,
                offsets      = offsets1),
           list(X            = X,
                coefficients = coeffs2,
                offsets      = offsets2))
head(p_mis)

## Example 2: will have two unique probabilities
unique(BDSE:::calc_missing_probabilities(
         list(X            = X,
              coefficients = c(0.5, 1, 0)),
         list(X            = X,
              coefficients = c(0.1, 0.5, 0))))
}
