% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_probs}
\alias{calc_probs}
\title{Calculate logistic probabilities}
\usage{
calc_probs(X, coefs, offsets = rep(0, nrow(X)))
}
\arguments{
\item{X}{A design matrix.}

\item{coefs}{A vector of coefficients.}

\item{offsets}{A numeric vector of offsets for the linear component; defaults
to zeros.}
}
\value{
A vector of logistic probabilities with length equal to number of
  rows in the design matrix.
}
\description{
Calculates the logistic regression probabilities given a design matrix,
coefficients and offsets. This is a wrapper for \code{\link[BDSE]{invlogit}}.
}
\examples{
dataset   <- data.frame(x = sample(10L, 20, replace = TRUE))
n_clus    <- 3L
cluster   <- sample(n_clus, nrow(dataset), replace = TRUE)
clus_effs <- runif(n_clus, -2, 2)

calc_probs(X       = model.matrix(~ x, dataset),
           coefs   = c(0.5, 0.25),
           offsets = clus_effs[cluster])

calc_probs(X     = model.matrix(~ x, dataset),
           coefs = c(0.5, 0.25))

}
