% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_in_list}
\alias{check_in_list}
\title{Check partial coverage input}
\usage{
check_in_list(has_partial_coverage, dataset)
}
\arguments{
\item{has_partial_coverage}{A Boolean indicating whether the input dataset
includes an indicator for partial coverage by design (defaults to
\code{FALSE}).}
}
\description{
Internal function to check for consistency of (List 1) partial coverage
information.
}
\examples{
BDSE:::check_in_list(FALSE, data.frame(X = 10))
BDSE:::check_in_list(TRUE,  data.frame(X = 10))
BDSE:::check_in_list(FALSE, data.frame(X = 10, in_list = T))
BDSE:::check_in_list(TRUE,  data.frame(X = 10, in_list = T))
 
}
