% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_in_target}
\alias{check_in_target}
\title{Check overcoverage input}
\usage{
check_in_target(has_overcoverage, dataset)
}
\arguments{
\item{has_overcoverage}{A Boolean indicating whether the input dataset
includes records of individuals not in the target dataset (defaults to
\code{FALSE}).}
}
\description{
Internal function to check for consistency of (List 2) overcoverage
information.
}
\examples{
BDSE:::check_in_target(FALSE, data.frame(X = 10))
BDSE:::check_in_target(TRUE,  data.frame(X = 10))
BDSE:::check_in_target(FALSE, data.frame(X = 10, in_target = T))
BDSE:::check_in_target(TRUE,  data.frame(X = 10, in_target = T))
 
}
