% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{colMedians}
\alias{colMedians}
\alias{rowMedians}
\title{Calculate medians}
\usage{
colMedians(x)

rowMedians(x)
}
\arguments{
\item{x}{A numeric matrix or data.frame with numeric columns.}
}
\value{
A vector of medians.
}
\description{
Calculates the column/row medians of a matrix or data.frame.
}
\examples{
colMedians(matrix(1:20, ncol = 2))
colMedians(data.frame(a = 1:10, b = 11:20))

rowMedians(matrix(1:20, ncol = 2))
rowMedians(data.frame(a = 1:10, b = 11:20))
}
