% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{color_by}
\alias{color_by}
\title{Make colors for plots}
\usage{
color_by(index_var, luminance = 70, chroma = 100)
}
\arguments{
\item{index_var}{A vector to color by.}
}
\value{
A vector of colors grouped by \code{index_var}.
}
\description{
Make colors for plots
}
