% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{colQuantiles}
\alias{colQuantiles}
\alias{rowQuantiles}
\title{Calculate quantiles}
\usage{
colQuantiles(x, p)

rowQuantiles(x, p)
}
\arguments{
\item{x}{A numeric matrix or data.frame with numeric columns.}
}
\value{
A vector of quantiles.
}
\description{
Calculates the column/row quantiles of a matrix or data.frame.
}
\examples{
colMedians(matrix(1:20, ncol = 2))
colQuantiles(matrix(1:20, ncol = 2), p = 0.5)
colMedians(data.frame(a = 1:10, b = 11:20))
colQuantiles(data.frame(a = 1:10, b = 11:20), p = 0.5)

rowMedians(matrix(1:20, ncol = 2))
rowQuantiles(matrix(1:20, ncol = 2), p = 0.5)
rowMedians(data.frame(a = 1:10, b = 11:20))
rowQuantiles(data.frame(a = 1:10, b = 11:20), p = 0.5)
}
