% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{convert_to_array}
\alias{convert_to_array}
\title{Convert list of matrices to single 3D array}
\usage{
convert_to_array(results_list, n_burnin = 0, n_thin = 1)
}
\arguments{
\item{results_list}{A list of matrices (sampler outputs), where each element
of the list corresponds to a chain; a result of a call to
\code{\link[BDSE]{bind_chains}}.}

\item{n_burnin}{An integer specifying the number of initial iterations to
discard from each chain.}

\item{n_thin}{An integer specifying the frequency of post-burnin iterations
to keep from each chain.}
}
\value{
A 3-dimensional array of thinned post-burn-in posterios samples, with
  dimensions iterations, chains, and sample.
}
\description{
Convert list of matrices to single 3D array
}
