% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Convert list of matrices to single data.frame}
\usage{
convert_to_df(results_list, n_burnin = 0, n_thin = 1)
}
\arguments{
\item{results_list}{A list of matrices (sampler outputs), where each element
of the list corresponds to a chain; a result of a call to
\code{\link[BDSE]{bind_chains}}.}

\item{n_burnin}{An integer specifying the number of initial iterations to
discard from each chain.}

\item{n_thin}{An integer specifying the frequency of post-burnin iterations
to keep from each chain.}
}
\value{
A data.frame of thinned post-burn-in posterior samples.
}
\description{
Convert list of matrices to single data.frame
}
