% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{Make distribution over unique values}
\usage{
distribute(df, contract_cols, complete_cols = names(df), as_dataframe = TRUE)
}
\arguments{
\item{df}{A data.frame or data.table.}

\item{contract_cols}{A character vector giving columns to aggregate.}

\item{complete_cols}{A character vector giving the columns in \code{df} to
keep which must include columns specified in \code{contract_cols}; defaults
to \code{names(df)}.}

\item{as_dataframe}{A logical scalar, specifying whether the output should be
coerced to a data.frame (defaults to \code{TRUE}). If \code{FALSE}, the
function returns a data.table.}
}
\value{
A data.frame or data.table of aggragted counts.
}
\description{
Aggregates count columns over unique values of the specified grouping
columns.
}
\details{
When the count column is 1 for all rows, this is the frequency distribution
over the unique values of the grouping variables.
}
\examples{
n   <- 20
counts <- sample(0:1, n, replace = TRUE)
df  <- data.frame(x1 = sample(1:2, n, replace = TRUE),
                  x2 = letters[1:2],
                  y1 = counts,
                  y2 = 1 - counts)

distribute(df, c("y1", "y2"))
distribute(df, c("y1"), c("x2", "y1"))

}
