% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_coefficients.R
\name{draw_coefficients}
\alias{draw_coefficients}
\title{Draw from the posterior distribution of individual-level effects}
\usage{
draw_coefficients(
  current_coefficients,
  prior_mean,
  prior_cov,
  proposal_cov,
  X,
  y,
  offsets = 0,
  weights = 1,
  counts = 1,
  proposal_scale = 1
)
}
\arguments{
\item{current_coefficients}{A numeric vector giving current coefficient
values (individual-level effects).}

\item{prior_mean}{A vector giving the mean for the MVN prior distribution
on the coefficients.}

\item{prior_cov}{A matrix giving the covariance for the MVN prior
distribution on the coefficients.}

\item{proposal_cov}{A matrix giving the covariance for the MVN proposal
distribution, with mean vector \code{current_coefficients}.}

\item{X}{A numeric (design) matrix with dimension \eqn{n} x \eqn{p}, where
\code{n} is the number of individuals, and \eqn{p} is the number of
covariates (including the intercept).}

\item{y}{A binary (0, 1) vector of length \eqn{n}, representing observations
of the response variable.}

\item{offsets}{An optional numeric vector of length \eqn{n} specifying the
values added to the linear component to account for effects not dependent
on \code{X}.}

\item{weights}{An optional numeric vector of length \eqn{n} specifying the
sample weights.}

\item{counts}{A numeric vector of counts of the number of observations in
each corresponding row of the design matrix \code{X}.}

\item{proposal_scale}{A numeric value to scale the covariance matrix
\code{proposal_cov} of the proporsal distribution.}
}
\value{
A named list, with the following elements:
\itemize{
    \item \code{coefficients}: the vector of sampled coefficients;
    \item \code{log_prior}: the prior log-likelihood of the sampled
            coefficient values;
    \item \code{log_likelihood}: the log-likelihood of the data model for
            the sampled coefficient values; and
    \item \code{was_updated}: a logical indicator of whether the coefficients
            vector was updated in the Metropolis-Hastings run.
}
}
\description{
This function uses the Metropolis-Hastings algorithm with the symmetric MVN
distribution
\preformatted{
  MVN(current_coefficients, proposal_cov),
}
to draw from the posterior distribution of the individual-level effects of
a logistic regression model. The prior distribution, with known mean and
covariance, is given by
\preformatted{
  MVN(prior_mean, prior_cov).
}
}
\details{
This function is a single iteration of the Metropolis-Hastings algorithm,
and is intended for use within a Gibbs sampler.
}
\examples{
n               <- 1000  # number of observations
n_clusters      <- 50    # number of clusters
clusters        <- sample(n_clusters, n, replace = TRUE)
X               <- cbind('(Intercept)' = rep(1, n),
                         age           = runif(n, 1, 85))
coeffs          <- c('(Intercept)' = 0.5, 
                     age           = 0.05)
y               <- rbinom(n, 1, 0.7)
cluster_effects <- rnorm(n_clusters, 0, 2)
offsets         <- cluster_effects[clusters]

draw_coefficients(current_coefficients = coeffs,
                  prior_mean           = rep(0, times = 2),
                  prior_cov            = diag(5^2, nrow = 2),
                  proposal_cov         = matrix(c(1.5^2, 1, 1, 1.5^2), nrow = 2),
                  X                    = X,
                  y                    = y,
                  offsets              = offsets)

}
