% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_Gibbs_samplers.R
\name{draw_initial_vals}
\alias{draw_initial_vals}
\title{Draw initial parameter values for the Gibbs sampler}
\usage{
draw_initial_vals(
  mean_coef1,
  mean_coef2,
  cov_coef1,
  cov_coef2,
  n_clusters = NULL,
  sd_min = NULL,
  sd_max = NULL,
  sd1_min = sd_min,
  sd1_max = sd_max,
  sd2_min = sd_min,
  sd2_max = sd_max
)
}
\arguments{
\item{mean_coef1, mean_coef2}{Each a numeric vector specifying the mean of the
normal distribution from which the coefficients for the corresponding list
is drawn.}

\item{cov_coef1, cov_coef2}{Each a matrix specifying the covariance of the
normal distribution from which the coefficients for the corresponding list
is drawn.}

\item{n_clusters}{An integer representing the number of clusters in the data.}

\item{sd_min, sd_max}{The lower and upper bound of the uniform distribution
from which the standard deviation of the distribution of the cluster
effects is drawn.}

\item{sd1_min, sd1_max, sd2_min, sd2_max}{The lower and upper bounds of the
uniform distributions from which the standard deviations of the
distributions of the cluster effects are drawn, when they differ for the
two lists.}
}
\value{
A named list of initial parameter values to be passed to
  \code{\link{initialise_output_storage}} through
  \code{run_Gibbs_sampler_*} functions, with elements:
  \code{coefficients1},
  \code{coefficients2},
  \code{cluster_sd1},
  \code{cluster_sd2},
  \code{cluster_effects1}, and
  \code{cluster_effects2}.
}
\description{
This function draws random initial parameter values for the Gibbs samplers.
The inidividual-level coefficients and the group-level cluster effects are
drawn from normal distributions, while the standard deviations of the cluster
effects distributions are drawn from uniform distributions.
}
\details{
The number of coeffcients drawn for the logistic regression coverage model
for each of the two lists is equal to the lengths of the mean vectors
\code{mean_coef1} and \code{mean_coef2}.

If the ranges of the standard deviations for the distribution of the cluster
effects are not the same for both lists, then they are specified using
\code{sd1_min}, \code{sd1_max}, \code{sd2_min}, and \code{sd2_max}. These
default to \code{sd_min} and \code{sd_max}.

The integers in the argument names refer to the indices of the two lists.
}
\examples{
q1 <- 3
q2 <- 5

## Non-hierarchical:
draw_initial_vals(mean_coef1 = rep(0, q1),
                  mean_coef2 = rep(0, q2),
                  cov_coef1  = diag(1.2, nrow = q1),
                  cov_coef2  = diag(0.8, nrow = q2))

## Hierarchical (2-level) model:
draw_initial_vals(mean_coef1 = rep(0, q1),
                  mean_coef2 = rep(0, q2),
                  cov_coef1  = diag(1.2, nrow = q1),
                  cov_coef2  = diag(0.8, nrow = q2),
                  n_clusters = 10,
                  sd_min     = 0.1,
                  sd_max     = 1)


}
