% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_normal_mean_and_sd.R
\name{draw_mean}
\alias{draw_mean}
\alias{draw_mean_normal}
\title{Draw from conditional posterior of the mean of a normal data model}
\usage{
draw_mean_normal(y, normal_sd, prior_mean, prior_sd, n = 1)
}
\arguments{
\item{y}{A numeric vector of observations (data).}

\item{normal_sd}{A numeric scalar specifying the standard deviation of the
normal model.}

\item{prior_mean}{A numeric scalar specifying the prior mean of the the mean
parameter of the normal distribution.}

\item{prior_sd}{A numeric scalar specifying the prior standard deviation of
the mean parameter of the normal distribution.}

\item{n}{An integer specifying the number of draws from the distribution.
(Defaults to 1.)}
}
\value{
A numeric vector of length \code{n} of sampled means.
}
\description{
This function draws samples from the posterior distribution of the mean
parameter of a normal model, given the variance parameter:
\deqn{
  Y \sim Normal(\mu, \sigma^2),
} 
for the conjugate normal prior on the mean parameter \eqn{\mu}:
\deqn{
  Normal(\mu_0, \sigma_0^2).
}
}
\examples{
draw_mean_normal(y          = rnorm(100, 3, 0.5),
                 normal_sd  = 2,
                 prior_mean = 0,
                 prior_sd   = 4,
                 n          = 5)
}
