% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_missing_rows.R
\name{draw_missing_counts}
\alias{draw_missing_counts}
\title{Draw missing counts}
\usage{
draw_missing_counts(
  X_obs1,
  X_obs2,
  coefficients1,
  coefficients2,
  theta,
  n_observed,
  n_target = NULL,
  offsets1 = 0,
  offsets2 = 0
)
}
\description{
Draw missing counts
}
