% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_missing_rows.R
\name{draw_missing_data_indices}
\alias{draw_missing_data_indices}
\title{Draw from the posterior predictive distribution of the missing records}
\usage{
draw_missing_data_indices(
  n_observed,
  X_obs1,
  X_obs2,
  coefficients1,
  coefficients2,
  theta,
  draw_n_target_posterior,
  n_target_posterior_args = NULL,
  offsets1 = 0,
  offsets2 = 0
)
}
\arguments{
\item{n_observed}{An integer specifying the number of rows in the observed
dataset.}

\item{X_obs1, X_obs2}{Each a design matrix produced by applying the coverage
model for the corresponding list (List 1 or List 2) to the observed dataset
(\eqn{D_{obs}}).}

\item{coefficients1, coefficients2}{Each a vector of individual-level effects
for the logistic regression coverage model for the corresponding list (List
1 or List 2).}

\item{theta}{A vector of probabilities (summing to 1) for the covariate
distribution.}

\item{draw_n_target_posterior}{A function to make a draw from the posterior
distribution of the target population size, given the size of the observed
portion of the population.}

\item{n_target_posterior_args}{Arguments for the function
\code{draw_n_target_posterior}.}

\item{offsets1, offsets2}{Each a numeric vector of offsets in the coverage
probbaility for the corresponding list; defaults to 0.}
}
\value{
A vector of indices of the sampled records (to build the unobserved
  or missing dataset).
}
\description{
This function draws samples from the posterior predictive distribution of the
covariate values of the unobserved or missing portion of the population,
given the data on the observed portion of the population.
}
\details{
The target population dataset is \eqn{[D_{obs}, D_{mis}]}, where
\eqn{D_{obs}} is the observed portion (captured by at least one of the two
lists List 1 and List 2), and \eqn{D_{mis}} is the unobserved or missing
portion of the dataset (not captured by either list). The corresponding
covariates matrices are \eqn{X_{obs}} and \eqn{X_{mis}}, which together form
the covariates matrix of the target population \eqn{X = [X_{obs}, X_{mis}]}.

To draw records (covariates vectors) for individuals not captures by the
lists, \code{draw_missing_data_indices} samples from the observed covariates
(\eqn{X_{obs}}), with probabilities adjusted using the coverage models. The
size of the target population, which determines the number of missing records
to draw, is drawn from its posterior distribution within
\code{draw_missing_data_indices}.

The two design matrices \code{X_obs1} and \code{X_obs2} have the same number
of rows, but may have different columns (depending on the terms in their
corresponding logistic regression coverage models).
}
\examples{
n_obs         <- 250  # number of observed data
n_clusters    <- 50   # number of clusters

X_obs         <- cbind('(Intercept)' = rep(1, n_obs),
                       sexM          = rbinom(n_obs, 1, 0.45),
                       age           = runif(n_obs, 1, 85))

clusters_obs  <- sample(n_clusters, n_obs, replace = TRUE)

theta         <- 1 / (5 * n_obs) * (rep(1, n_obs) +
                                    rmultinom(1,
                                              size = 4 * n_obs,
                                              prob = rep(1 / n_obs, n_obs)))

cluster_effects1 <- rnorm(n_clusters, 0, 0.5)
cluster_effects2 <- rnorm(n_clusters, 0, 0.5)

offsets1 <- cluster_effects1[clusters_obs]
offsets2 <- cluster_effects2[clusters_obs]

sum(theta)   # must sum to 1

X_mis_indices <-
  draw_missing_data_indices(n_observed         = n_obs,
                            X_obs1             = X_obs,
                            X_obs2             = X_obs,
                            coefficients1      = c(0.5, -0.5, 0.01),
                            coefficients2      = c(0.5, 0, 0.02),
                            theta              = theta,
                            draw_n_target_posterior = draw_n_target_jeffreys,
                            n_target_posterior_args = NULL,
                            offsets1   = offsets1,
                            offsets2   = offsets2)

length(X_mis_indices) + n_obs  # size of target population

X_mis <- X_obs[X_mis_indices, ]
head(X_mis)

}
