% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_region_effects.R
\name{draw_region_effects}
\alias{draw_region_effects}
\title{Draw region effects from its conditional posterior distribution}
\usage{
draw_region_effects(
  cluster_effects,
  regions,
  n_regions,
  clusters_sd,
  prior_mean,
  prior_sd,
  offsets = 0
)
}
\arguments{
\item{cluster_effects}{A numeric vector of low-level cluster effects.}

\item{regions}{An integer or numeric vector, the same length as
\code{cluster_effects}, specifying the indices of the regions to which the
clusters corresponding to \code{cluster_effects} belong.}

\item{n_regions}{An integer specifying the number of regions in the
population.}

\item{clusters_sd}{A numeric scalar representing the between-cluster standard
deviation.}

\item{prior_mean}{A numeric vector of length \code{n_regions} specifying the
mean of the prior distribution of the region-level effects.}

\item{prior_sd}{A numeric scalar specifying the standard deviation of the
prior distribution of the region-level effects.}

\item{offsets}{An optional numeric vector, having the same length as
\code{cluster_effects}, of known offsets (defaults to 0); see Details.}
}
\value{
A numeric vector of length \code{n_regions} of the drawn sample of 
  region effects.
}
\description{
This function draws a vector of high-level region effects from its
conditional posterior distribution, given a vector of low-level cluster
effects. (In this setting, each region contains one or more clusters).
}
\details{
This function generates a draw from the conditional posterior of the means of
the normal data model with known standard deviation:
\deqn{
   Y_j ~ Normal(\mu_{reg(j)} + \eta_j, \sigma^2),
 }
where \eqn{Y_j} is the \eqn{j}th cluster effect, \eqn{\mu_{reg(j)}} is the
effect of the region to which cluster \eqn{j} belongs, and \eqn{\eta_j} is
some known offset for cluster \eqn{j} (e.g. a linear combination of
cluster-level covariates).

The conjugate prior distribution of the region effects 
(\eqn{\mu_1, \dots, \mu_{n_regions}}) is 
\preformatted{
  Normal(prior_mean, prior_sd).
}
}
\examples{
n_clusters      <- 1000     # number of clusters
n_regions       <- 20       # number of regions

# draw region indices (allow unobserved regions) 
regions         <- sample(sample(n_regions, floor(0.8 * n_regions)),
                          size = n_clusters,
                          replace = TRUE)
table(regions)

cluster_effects <- rnorm(n_clusters, 2, 1)
prior_mean      <- rnorm(n_regions)
prior_sd        <- 3
clusters_sd     <- 1.2

draw_region_effects(cluster_effects = cluster_effects,
                    regions         = regions,
                    n_regions       = n_regions,
                    clusters_sd     = clusters_sd,
                    prior_mean      = prior_mean,
                    prior_sd        = prior_sd)

}
