% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_normal_mean_and_sd.R
\name{draw_sd}
\alias{draw_sd}
\alias{draw_sd_uniform}
\alias{draw_sd_invchisq}
\title{Sample from posterior distribution of the standard deviation parameter of a 
Normal model}
\usage{
draw_sd_uniform(residuals, prior_lower, prior_upper, n = 1)

draw_sd_invchisq(residuals, prior_alpha, prior_beta, n = 1)
}
\arguments{
\item{residuals}{A vector of residuals \eqn{Y_i - \mu_i}.}

\item{prior_lower, prior_upper}{Numeric values specifying the lower and upper 
bounds for the uniform prior on \eqn{\sigma}.}

\item{n}{An integer representing the number of posterior samples to draw 
(defaults to 1).}

\item{prior_alpha, prior_beta}{Numeric values representing the shape 
(\code{alpha}) and scale (\code{beta}) of the inverse Chi-squared prior 
distribution on the variance \eqn{\sigma^2}.}
}
\description{
Generate draws from the posterior for the standard deviation of the 
normal model with known mean(s):
\deqn{
   Y_i ~ Normal(\mu_i, \sigma^2) .
 }
This function is used to update the standard deviation parameter of the 
normal prior distribution of the cluster effects.
}
\details{
The draws are made from the posterior distribution of either the precision 
\eqn{\tau^2 = 1 / \sigma^2}, or the variance \eqn{\sigma^2}, and then 
transformed to get the standard deviation.

\code{draw_sd_uniform} returns posterior draws of \eqn{\sigma} when 
the prior is \eqn{\sigma ~ U(prior_lower, prior_upper)}. The draws are made 
from the posterior distribution of the precision, which is the Gamma 
distribution
\deqn{
   Gamma ( (m - 1) / 2, \sum(Y_i - \mu_i)^2 / 2 ),
}
truncated to include only \eqn{\sigma \in (prior_lower, prior_upper)}, 
where \eqn{m} is the number of residuals. The drawn precisions are then 
transformed to get the standard deviations: \eqn{\sigma = 1 / \sqrt \tau^2}.

\code{draw_sd_invchisq} returns posterior draws of \eqn{\sigma} when 
the (conjugate) prior on the variance is 
\eqn{\sigma^2 ~ Scaled Inv-\chi^2(\alpha, \beta)}, where \eqn{\alpha} 
(\code{prior_alpha}) is the shape and \eqn{\beta} (prior_beta) is the scale. 
The posterior distribution of the variance parameter is
\deqn{
  Scaled Inv-\chi^2( m + \alpha, \alpha * \beta * \sum(Y_i - \mu_i)^2 / 
    (m + \alpha) ) ,
}
where \eqn{m} is the number of residuals.
}
\examples{
draw_sd_uniform(residuals   = rnorm(130, 0, 2),
                prior_lower = 1,
                prior_upper = 5)

draw_sd_invchisq(residuals   = rnorm(130, 0, 2),
                 prior_alpha = 2,
                 prior_beta  = 1,
                 n = 2)

}
