% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{extract_coefs}
\alias{extract_coefs}
\title{Extract Gibbs posterior sample for coefficients}
\usage{
extract_coefs(list_id, results, true_names)
}
\arguments{
\item{list_id}{An integer, the list ID (1 or 2).}

\item{results}{A data.frame of Gibbs sampler results (the output of
\code{\link[BDSE]{convert_to_df}}).}

\item{true_names}{A character vector of names of the coefficients appearing
in the order needed.}
}
\value{
A data.frame of posterior samples.
}
\description{
Extracts the Gibbs sampler samples for the coefficients of the given list ID.
}
