% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertility}
\alias{fertility}
\title{Fertility data}
\format{
A data.frame with 11 rows and 3 variables:
\describe{
  \item{mothers_age_grp}{...}
  \item{average_births}{...}
  \item{fert_rate}{...}
}
}
\usage{
fertility
}
\description{
A dataset containing the fertility rates for women of childbearing age by age
group.
}
\keyword{datasets}
