% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{first}
\alias{first}
\title{Restrict array}
\usage{
first(arr, index = 1, dim = 1)
}
\arguments{
\item{arr}{Array to restrict in one dimension.}

\item{index}{Index along dimension \code{dim} to restrict to.}

\item{dim}{Dimension to collapse.}
}
\value{
An array \code{arr[, ..., index, ...]} where the \code{index} value
  occurs in the \code{dim} place.
}
\description{
Restrict array
}
\examples{
  arr <- array(sample(360, 360),
               dim = 3:6)
  identical(arr[1, , , ], BDSE:::first(arr))
}
