% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_dataset.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Draw records from a synthetic base dataset to create a larger dataset}
\usage{
generate_dataset(n, base_dataset, weights = NULL, seed = NULL)
}
\arguments{
\item{n}{An integer specifying the size of the dataset to generate.}

\item{base_dataset}{A data.frame of covariates.}

\item{weights}{A numeric vector of sampling weights, the same length as the
number of rows in the base dataset. (Defaults to \code{NULL}.)}

\item{seed}{An integer specifying the seed for the random number generator.
(Defaults to \code{NULL}.)}
}
\description{
Draw records from a synthetic base dataset to create a larger dataset
}
\examples{
base_dataset <- expand.grid(sex       = c("M", "F"),
                            income    = seq(0, 50000, 10000),
                            age_grp   = as.character(1:5),
                            ethnicity = LETTERS[1:5])

synthetic_dataset <- generate_dataset(n            = 1000,
                                      base_dataset = base_dataset,
                                      seed         = 234)

str(synthetic_dataset)

}
