% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_dataset.R
\name{generate_target_dataset}
\alias{generate_target_dataset}
\title{Draw records from a synthetic base dataset to create a target dataset}
\usage{
generate_target_dataset(n_target, base_dataset, weights = NULL, seed = NULL)
}
\arguments{
\item{n_target}{An integer specifying the size of the target population.}

\item{base_dataset}{A data.frame of covariates.}

\item{weights}{A numeric vector of sampling weights, the same length as the
number of rows in the base dataset. (Defaults to \code{NULL}.)}

\item{seed}{An integer specifying the seed for the random number generator.
(Defaults to \code{NULL}.)}
}
\value{
A data.frame of covariates with \code{n_target} records.
}
\description{
This function generates a target (population) dataset by randomly sampling
from a given base dataset.
}
\examples{
base_dataset <- expand.grid(sex       = c("M", "F"),
                            income    = seq(0, 50000, 10000),
                            age_grp   = as.character(1:5),
                            ethnicity = LETTERS[1:5])

target_dataset <- generate_target_dataset(n_target     = 10000,
                                          base_dataset = base_dataset,
                                          seed         = 234)

str(target_dataset)

}
