% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{get_coefs_idx}
\alias{get_coefs_idx}
\title{Get positions of coefficients}
\usage{
get_coefs_idx(list_id, true_names, pars)
}
\arguments{
\item{list_id}{An integer, the list ID (1 or 2).}

\item{true_names}{A character vector of names of the coefficients appearing
in the order needed.}

\item{pars}{A character vector of Gibbs sample coefficient/parameter names to
search in.}
}
\value{
An integer vector of indices.
}
\description{
Gets the positions of the coefficients from the Gibbs sampelr results for the
given list.
}
