% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit function}
\usage{
invlogit(x, simplify = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}
}
\value{
A vector of probabilities.
}
\description{
Given a real value, returns the inverse-logit or logistic function value.
}
\details{
The logistic function \eqn{1 / (1 + \exp(-x))} is a mapping from
\eqn{(-\Infty, \Infty)} to \eqn{(0, 1)}.

If input is not a vector (e.g. is a matrix), function produces a warning,
while returning a vector.
}
\examples{
BDSE:::invlogit(matrix(c(1:4), ncol = 2))

## warns if input is coerced to vector
BDSE:::invlogit(matrix(c(1:4), ncol = 1))
BDSE:::invlogit(matrix(c(1:4), nrow = 1))

}
