% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{log_dnorm}
\alias{log_dnorm}
\title{Logarithm of the Normal density function}
\usage{
log_dnorm(x, mean, sd)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{mean}{A vector of means.}

\item{sd}{A vector of standard deviations.}
}
\value{
A vector of log densities.
}
\description{
Calculates the natural logarithm of a vector of normal density values.
}
