% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictors.R
\name{make_basis_matrix}
\alias{make_basis_matrix}
\alias{make_basis_matrix.bs}
\alias{make_basis_matrix.ti}
\title{Create a basis matrix for a polynomial spline}
\usage{
make_basis_matrix(
  variables,
  dataset,
  base_name,
  smooth_type,
  params = NULL,
  include_mains = TRUE
)

make_basis_matrix.bs(variable, dataset, base_name, bs_params)

make_basis_matrix.ti(
  variables,
  dataset,
  base_name,
  ti_params,
  include_mains = FALSE
)
}
\arguments{
\item{base_name}{A string specifying the name of the variable to use as
the base name for the columns of the basis matrix. (For example, for "age"
and 10 terms, the column names will be "age1", "age2", ..., "age10".)}

\item{variable}{A numeric vector: the predictor variable the basis matrix is
created from.}

\item{bs_params}{A named list of parameters passed to
\code{\link[splines]{bs}}.}
}
\value{
A matrix.
}
\description{
This function is a wrapper for \code{\link[splines]{bs}}, and is used to
allow specifying the piecewise polynomial terms of the associated predictor
variable in the logistic regression coverage model.
}
\examples{
dataset <- data.frame(sex = sample(2, 1000, replace = TRUE),
                      age = sample(100, 1000, replace = TRUE))

bs_matrix <- make_basis_matrix(variable    = "age", 
                               dataset     = dataset,
                               base_name   = "age_bs",
                               smooth_type = "bs",
                               params      = list(knots = seq(15, 85, 10)))

head(bs_matrix, 10)

}
