% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_dataset.R
\name{make_clusters_dataset}
\alias{make_clusters_dataset}
\title{Make cluster-level dataset}
\usage{
make_clusters_dataset(
  target_dataset,
  obs_dataset,
  joining_vars,
  target_vars = NULL,
  obs_vars = NULL,
  target_clusters_col = "cluster",
  obs_clusters_col = "cluster_obs"
)
}
\arguments{
\item{target_dataset, obs_dataset}{Each s data.frame, the target or observed
dataset including the clustering variables.}

\item{joining_vars}{A character vector specifying the clustering variables on
which to join the target and observed datasets.}

\item{target_vars, obs_vars}{Each a character vector specifying any additional
target or observed dataset columns to be included in the output. (Both
default to \code{NULL}.)}

\item{target_clusters_col, obs_clusters_col}{Each a string specifying the
sequential cluster IDs column in the target or observed dataset. (Defaults
to "cluster" for the target dataset and "cluster_obs" for the observed
dataset.)}
}
\value{
A data.frame, the merged target and observed cluster-level datasets.
  The observed variable columns will be \code{NA} for all clusters in the
  target dataset that were not observed.
}
\description{
Makes the cluster-level (area-level) dataset indicating observed and
unobserved clusters.
}
