% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_count}
\alias{make_count}
\title{Make counts vector}
\usage{
make_count(n, rows)
}
\arguments{
\item{n}{Number of elements.}

\item{rows}{Indices of repeated elements.}
}
\value{
Counts of the number of occurences of each element.
}
\description{
Make counts vector
}
\examples{
BDSE:::make_count(10, c(1, 1, 2, 1, 7))
}
