% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_summary.R
\name{make_coverage_summary}
\alias{make_coverage_summary}
\title{Summerise coverage}
\usage{
make_coverage_summary(dataset)
}
\arguments{
\item{dataset}{A data.frame with cell and list inclusion indicators.}
}
\value{
A named list with four elements: 
  \code{out_target_summary} (if dataset has list overcoverage);
  \code{in_target_coverage_summary};
  \code{in_target_cell_counts}; and
  \code{in_target_cell_proportions}.
}
\description{
This function creates a summary of list coverage for a given dataset.
}
\details{
The dataset must include the list and cell indicator columns \code{L1},
\code{L2}, \code{Y11}, \code{Y10}, \code{Y01}, and \code{Y00}. In the
presence of overcoverage, it must also include the target inclusion indicator
column \code{in_target}.
}
\examples{
n_union  <- 1200
n_target <- 1000

L1 <- sample(c(TRUE, FALSE), n_target, prob = c(0.6, 0.4), replace = TRUE)
L2 <- sample(c(TRUE, FALSE), n_target, prob = c(0.8, 0.2), replace = TRUE)

## Dataset without overcoverage:
target_dataset <- data.frame(L1  = L1,
                             L2  = L2,
                             Y11 = L1 & L2,
                             Y10 = L1 & !L2,
                             Y01 = !L1 & L2,
                             Y00 = !L1 & !L2)

make_coverage_summary(target_dataset)

## Dataset with overcoverage:
union_dataset <- rbind(target_dataset,
                       data.frame(L1  = rep(FALSE, n_union - n_target),
                                  L2  = TRUE,
                                  Y11 = NA,
                                  Y10 = NA,
                                  Y01 = NA,
                                  Y00 = NA))

union_dataset[["in_target"]] <- c(rep(TRUE, n_target), 
                                 rep(FALSE, n_union - n_target))
                                  
make_coverage_summary(union_dataset)

}
