% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictors.R
\name{make_ethnicity_indicators}
\alias{make_ethnicity_indicators}
\title{Make ethnicity indicators}
\usage{
make_ethnicity_indicators(ethnicity, single_ethnicities = c(1:6, 9))
}
\arguments{
\item{ethnicity}{A vector of numeric ethnicity codes.}

\item{single_ethnicities}{A vector specifying the values of the
unique single-ethnicity levels. (Defaults to the single ethnicity codes
\code{c(1:6, 9)}.)}
}
\value{
A data.frame of ethnicity indicators.
}
\description{
This function converts the vector of ethnicity codes into a data.frame of
ethnicity indicators for single ethnicities.
}
\details{
In \code{base_dataset}, the \code{ethnicity} column is the integer-valued
ethnicity codes, where single digits refer to single-ethnicity groups and 
double digits refer to bi-ethnicity groups. This function converts this
ethnicity encoding to a data.frame of single-ethnicity indicators, where each
individual (row) has at least one single-ethnicity indicator set to 1.
}
\examples{
dataset <- BDSE::base_dataset
table(dataset$ethnicity)

ethnicity_df <- make_ethnicity_indicators(dataset[["ethnicity"]])

head(ethnicity_df)
colSums(ethnicity_df)

## Checks:
all(rowSums(ethnicity_df) > 0)
all(ethnicity_df[ethnicity_df[, 7] == 1, -7] == 0)  # 7 = ethnicity9 = "other"

}
