% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_predictors.R
\name{make_ethnicity_interactions}
\alias{make_ethnicity_interactions}
\title{Make ethnicity interactions}
\usage{
make_ethnicity_interactions(
  dataset,
  interactions = list(list(1, 2:6), list(2, 3:6), list(3, 4:6), list(4, 5:6), list(5, 6)),
  base_name = "ethnicity"
)
}
\arguments{
\item{dataset}{A data.frame with the single ethnicity indicators as columns.}

\item{interactions}{A list, each element of which is a 2-element list with
its first element the code for the single indicator, and its second
element, the codes for the single ethnicities it interacts with. (See
defaults.)}

\item{base_name}{A string represnting the base name of the variable (defaults
to \code{"ethnicity"}).}
}
\value{
A data.frame of ethnicity interactions indicators.
}
\description{
This functions creates interaction indicators from the single-ethnicity 
indicators (which are assumed to exist in the dataset).
}
\examples{
dataset <- BDSE::base_dataset
dataset <- cbind(dataset, 
                 make_ethnicity_indicators(dataset[["ethnicity"]]))

ethnicity_interactions_df <- make_ethnicity_interactions(dataset = dataset)

head(ethnicity_interactions_df)

}
