% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_file_path}
\alias{make_file_path}
\title{Function to make Gibbs sampler result file path}
\usage{
make_file_path(path, model, date_time, seed)
}
\arguments{
\item{path}{A string specifying the path where the file must be stored.}

\item{model}{A string specifying the model fit in the sampler; options are
"L1", "L2a" or "L2b".}

\item{date_time}{A string specifying the date and time the sample is run.}

\item{seed}{An integer specifying the random seed used in the sampler.}
}
\value{
A string, the results file path.
}
\description{
This (internal) function makes the file path for the Gibbs samplers.
}
\examples{
BDSE:::make_file_path("dir/", "L2a", "20120131", 123)
BDSE:::make_file_path("dir", "L2a", "20120131", 123)
}
