% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{make_results_list}
\alias{make_results_list}
\title{Function to import Gibbs sampler result files}
\usage{
make_results_list(directory, file_names)
}
\arguments{
\item{directory}{A string specifying the directory in which the result files
are stored.}

\item{file_names}{A character vector of file names (with extensions, e.g.
"gibbs_chain1.RDS")}
}
\value{
A list of length equal to the number of chains (files).
}
\description{
Given the directory and file names (corresponding to each chain), reads in	
the results files and merges them into a single list.
}
