% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_dataset_wrapper.R
\name{make_synthetic_data}
\alias{make_synthetic_data}
\title{A wrapper to generate all errors}
\usage{
make_synthetic_data(
  seed,
  dataset,
  coefficients1,
  coefficients2,
  formula1,
  formula2,
  clusters_sd1,
  clusters_sd2,
  clustering_var,
  formula_sens = NULL,
  formula_target = NULL,
  clusters_sd_target = NULL,
  coefficients_sens = NULL,
  coefficients_target = NULL,
  clustering_var_target = NULL,
  n_validated = 50000
)
}
\description{
TODO: Fill in parameter details and description later
}
