% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{normalize}
\alias{normalize}
\title{Min-max normalize a numeric vector}
\usage{
normalize(var, l_min = -1, l_max = 1)
}
\arguments{
\item{var}{A numeric vector.}

\item{l_min}{A scalar specifying the lower limit of the normalized range.}

\item{l_max}{A scalar specifying the upper limit of the normalized range.}
}
\value{
A numeric vector of normalized values.
}
\description{
Normalizes a numeric vector to within a specified range (by default
\eqn{[-1, 1}).
}
\examples{
(x <- sample(100, 10))
normalize(x)

}
