% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{order_by_name}
\alias{order_by_name}
\title{Order elements of a list or vector by name}
\usage{
order_by_name(input, as_integer = TRUE)
}
\arguments{
\item{input}{A named list or vector.}

\item{as_integer}{Logical, indicating whether the names must be coerced to
integers before ordering.}
}
\value{
The ordered input vector or list.
}
\description{
Orders elements of a named list or vector by element name.
}
\examples{
BDSE:::order_by_name(c('2' = 26.4, '1' = 9.5, '3' = 12))

BDSE:::order_by_name(list('2' = 26.4, '1' = 9.5, '3' = 12))
}
