% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{record_count}
\alias{record_count}
\title{Indicator function}
\usage{
record_count(burn_in, thinning, count, n_sim)
}
\arguments{
\item{burn_in}{Positive integer.}

\item{thinning}{Positive integer. If not provided can be calculated from
\code{count} and \code{nsim}.}

\item{count}{Positive integer. Only used if \code{thinning} is not provided.}

\item{n_sim}{Positive integer. Only used if \code{thinning} is not provided.}
}
\value{
A function \eqn{N \rightarrow N} that returns true for every
  \code{thinning}th value after the \code{burnin}.
}
\description{
Indicator function
}
