% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rtruncated_norm}
\alias{rtruncated_norm}
\title{Draw from a truncated normal}
\usage{
rtruncated_norm(n, mean, sd, min = -Inf, max = Inf, ntry = 1e+06)
}
\arguments{
\item{n}{An integer specifying the number of variates to generate.}

\item{mean}{A vector of means.}

\item{sd}{A vector of standard deviations.}

\item{min, max}{The truncation bounds; default to \code{-Inf} and \code{Inf}.}
}
\value{
A vector of length \code{n}.
}
\description{
Draw from a truncated normal
}
\examples{
BDSE:::rtruncated_norm(n    = 2, 
                       mean = c(0, 10), 
                       sd   = c(1, 2.5),
                       min  = c(-3, 5), 
                       max  = c(3, 25))

}
