% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Gibbs_sampler_1level.R
\name{run_Gibbs_sampler_1level}
\alias{run_Gibbs_sampler_1level}
\title{Gibbs sampler (single-level coverage model)}
\usage{
run_Gibbs_sampler_1level(
  seed,
  n_sims,
  obs_distribution,
  formula1,
  formula2,
  proposal_cov_coef1,
  proposal_cov_coef2,
  prior_mean_coef1,
  prior_mean_coef2,
  prior_cov_coef1,
  prior_cov_coef2,
  initial_vals = NULL,
  restart_from = NULL,
  proposal_scale_coef1 = 1,
  proposal_scale_coef2 = 1,
  func_n_target_posterior = list(func = BDSE:::draw_n_target_jeffreys, args = NULL),
  periodic_save = NULL,
  print_every = 0L,
  record_count = function(...) FALSE,
  return_obs_dist = FALSE
)
}
\arguments{
\item{seed}{An integer specifying the random seed.}

\item{n_sims}{An integer specifying the number of simulations (or runs of the
Gibbs sampler).}

\item{obs_distribution}{A data.frame of the unique observed unit records.}

\item{formula1, formula2}{Each a formula for creating the design matrix for
the coverage model of the corresponding list (list 1 or 2).}

\item{proposal_cov_coef1, proposal_cov_coef2}{Each a covariance matrix for the
multivariate normal proposal distribution used to draw from the conditional
posterior distribution of the coefficients in the coverage model for the
corresponding list (List 1 or 2).}

\item{prior_mean_coef1, prior_mean_coef2}{Each a vector of means for the
normal prior distribution of the coeffcients of the coverage model for the
corresponding list.}

\item{prior_cov_coef1, prior_cov_coef2}{Each a covariance matrix for the
normal prior distribution of the coeffcients of the coverage model for the
corresponding list.}

\item{initial_vals}{A named list with initial values for the relavent model
parameters: here, the coefficient vectors \code{coefficients1} and
\code{coefficients2}.}

\item{proposal_scale_coef1, proposal_scale_coef2}{Each a numeric scalar
specifying the scaling factor of the covariance matrix of the M-H proposal
distribution for the coefficients for the corresponding list.}

\item{func_n_target_posterior}{A named list with elements \code{func} and
\code{args} giving the function used to draw from the posterior distrbution
of the target size (\code{n_target}) and the arguments to the function.
(Defaults to the posterior for a Jeffreys prior.)}

\item{periodic_save}{A named list with two elemnts: \code{every}, an integer
\eqn{n} specifying that every \eqn{n}th iteration must be saved, and
\code{where}, a string specifying the directory where the results should be
saved. (Defaults to \code{NULL}.)}

\item{print_every}{An integer \eqn{n} specifying that after every \eqn{n}th
iteration a progress report should be printed (to console).}

\item{record_count}{A function indicating which counts (iterations) should be
stored in the output object; see \code{\link[BDSE]{record_count}}.}

\item{return_obs_dist}{Logical, indicating whether the observed frequency
distribution should be stored in the output object; defaults to
\code{TRUE}.}
}
\value{
A named list of parameter values generated by the Gibbs sampler: 
  \itemize{ 
    \item \code{n_target} A vector of sampled target population sizes;
    \item \code{coefficients1, coefficients2}: Matrices of sampled 
      coefficients for the coverage models of lists 1 and 2; 
    \item \code{update_coef1, update_coef2}: Each a vector indicating whether
      the coefficients vector for the corresponding list was updated at each
      iteration; 
    \item \code{log_like1, log_like2}: Each a vector of log-likelihoods of
      the coefficients vector for the corresponding list.
    \item \code{run_time}: The run-time of the Gibbs sampler.
  }
}
\description{
This function runs a Gibbs sampler to approximate the posterior predictive 
distribution of the unobserved portion of the target population, given the
portion of the population observed (in Lists 1 and 2).
}
\details{
The function assumes that, along with the covariates, the data.frame
\code{obs_distribution} includes columns of counts for the four cells, named
\code{Y11}, \code{Y10}, \code{Y01}, and \code{Y00}.
}
\examples{
n_obs  <- 100000 
n_sims <- 10L 

L     <- data.frame(L1 = c(1, 1, 0), 
                    L2 = c(1, 0, 1))[sample(3, n_obs, replace = TRUE), ]
                    
D_obs <- data.frame(age = sample(90, n_obs, replace = TRUE),
                    sex = sample(c("M", "F"), n_obs, replace = TRUE),
                    Y11 = as.integer(L[[1]] & L[[2]]),
                    Y10 = as.integer(L[[1]] & !L[[2]]),
                    Y01 = as.integer(!L[[1]] & L[[2]]),
                    Y00 = 0L)

obs_distribution <- BDSE::distribute(
                      D_obs,
                      contract_cols = c("Y11", "Y10", "Y01", "Y00"))

cover_formula <- ~ age + sex  # terms: Intercept, age, and sexM (or sexF)

initial_vals  <- list(coefficients1 = c(0.5, 0.001, 0.75),
                      coefficients2 = c(0.5, 0.001, 0.75))
 
n_coef1 <- length(initial_vals[["coefficients1"]])
n_coef2 <- length(initial_vals[["coefficients2"]])

Gibbs_sample <- run_Gibbs_sampler_1level(
                  seed               = 296,
                  n_sims             = n_sims,
                  obs_distribution   = obs_distribution,
                  formula1           = cover_formula,
                  formula2           = cover_formula,
                  proposal_cov_coef1 = diag(0.5, nrow = n_coef1),
                  proposal_cov_coef2 = diag(0.5, nrow = n_coef2),
                  prior_mean_coef1   = rep(0, n_coef1),
                  prior_mean_coef2   = rep(0, n_coef2),
                  prior_cov_coef1    = diag(100^2, nrow = n_coef1),
                  prior_cov_coef2    = diag(100^2, nrow = n_coef2),
                  initial_vals       = initial_vals,
                # periodic_save      = list(every = n_sims,
                #                           where = file.path(getwd(), "out")),
                  print_every        = 1L)

str(Gibbs_sample)

}
