% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_expj}
\alias{sample_expj}
\title{Weighted sample without replacement from table}
\usage{
sample_expj(n, size, counts, probs)
}
\arguments{
\item{n}{A positive integer; values are sampled from \code{1:n}.}

\item{size}{Positive integer specifying how many values to sample.}

\item{counts}{Numeric vector of length \code{n} giving the maximumum number
of each integer that can be drawn.}

\item{probs}{Numeric vector of length \code{n} giving weights for each
integer.}
}
\value{
A vector of values in \code{1:n} such that \code{i} does not appear
  more than \code{count[i]} times.
}
\description{
Weighted sample without replacement from table
}
\examples{
   sample_expj(20,
               50,
               sample(10, 20, replace = TRUE),
               round(runif(20), 2))

}
