% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_rank}
\alias{sample_rank}
\title{Weighted sample without replacement with repeated values}
\usage{
sample_rank(n, size, counts, probs)
}
\arguments{
\item{n}{Number of items to choose from.}

\item{size}{Number of items to choose.}

\item{counts}{Vector of length \code{n} giving the number of times each value
occurs.}

\item{probs}{Vector of weights giving relative likelihood of sampling each
value.}
}
\value{
A vector of length \code{size} with elements of \code{1:n}.
}
\description{
Weighted sample without replacement with repeated values
}
