% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_without_collisions}
\alias{sample_without_collisions}
\title{Sample without collision}
\usage{
sample_without_collisions(support, bad, prob = NULL)
}
\description{
Sample values from support excluding values in bad.
}
