% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions_processing_results.R
\name{summarize_single_chain}
\alias{summarize_single_chain}
\title{Makes summary of Gibbs sampler stored object}
\usage{
summarize_single_chain(result)
}
\arguments{
\item{result}{A list with named elements.}
}
\value{
A summary data.frame of names, dimensions, and classes of the input
  list.
}
\description{
Makes a summary data.frame of a Gibbs sampler output (single chain), i.e the	
stored object returned by the sampler function \code{run_Gibbs_sampler}.
}
