% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{t_apply}
\alias{t_apply}
\title{Fast tapply}
\usage{
t_apply(X, INDEX, FUN)
}
\arguments{
\item{X}{A numeric vector.}

\item{INDEX}{A vector of the same length as \code{X}.}

\item{FUN}{A function.}
}
\value{
\code{tapply(X, INDEX, FUN = FUN)}
}
\description{
Fast tapply
}
