% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{zip}
\alias{zip}
\title{Zip together lists}
\usage{
zip(...)
}
\arguments{
\item{...}{Vectors or lists, all of the same length.}
}
\value{
A list with its \eqn{i}th element being the list of the \eqn{i}th
  elements of the inputs.
}
\description{
Zip together lists
}
\examples{
  zip(x = 1:3, y = letters[1:3])

}
